removed all sub controllers
+ the views handle "get_command"
+ the views handle the commands
+ change ask_for_command for get_command
+ imports refactoring
+ removed unused commands
