+++ .gitignore
@@ -1,3 +1,4 @@
 .venv
 __*
 !__init__.py
+.*
