--- app/controllers/__init__.py
@@ -1,7 +1,0 @@
-# flake8: noqa: F401
-"""Controllers package."""
-
-from .abc import Controller
-from .main_page import MainPageController
-from .add_player import AddPlayerController
-from .new_tournament import NewTournamentController
