--- app/controllers/abc.py
@@ -1,28 +1,0 @@
-"""Controller interface."""
-
-from abc import ABC
-
-from app.commands import BlankCommand, QuitCommand, WrongCommand, NavigationCommand
-from app.views import View
-
-
-class Controller(ABC):
-    """Base controller."""
-
-    def __init__(self, *args, **kwargs):
-        """Init."""
-        self.commands = [QuitCommand, BlankCommand, NavigationCommand]
-        self.view: View = View(self.commands)
-
-    def get_command(self):
-        """Get the command."""
-        if self.view.type == "event":
-            return self.view.ask_for_command()
-
-        choice = input()
-
-        for Command in self.commands:
-            if choice in Command.get_choices():
-                return Command(choice)
-
-        return WrongCommand(choice)
