--- app/controllers/add_player.py
@@ -1,15 +1,0 @@
-"""Handle the main page."""
-
-from app.views import AddPlayerView
-
-from .abc import Controller
-
-
-class AddPlayerController(Controller):
-    """Main page."""
-
-    def __init__(self):
-        """Init."""
-        super().__init__()
-
-        self.view = AddPlayerView(self.commands)
